requirements.txt
langchain>=0.0.142
langchain_openai
langchain_groq
nicegui
fauna
selenium
pyyaml
bs4